requirements.txt
elasticsearch==7.9.1
psycopg2-binary==2.8.6
urllib3==1.25.11
